chore: features/statusディレクトリを整理する
