+++ src/features/status/components/partials/index.ts
@@ -1,0 +1,1 @@
+export * from "./HeadLine";
