+++ src/features/status/components/ui/LevelControl.tsx
