+++ src/features/status/components/ui/index.ts
@@ -1,3 +1,2 @@
 export * from "./LevelControl";
 export * from "./TotalEffortValue";
-export * from "./TableHead";
