+++ src/features/status/table/components/partials/index.ts
@@ -1,2 +1,1 @@
-export * from "./HeadLine";
 export * from "./TableBody";
