+++ src/features/status/table/components/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "./TableHead";
