+++ src/features/status/tableRow/components/StatusTableBody.tsx
@@ -1,36 +1,38 @@
 import { useState, useCallback, useEffect } from "react";
 
 import { currentEffortValueAtom, adjustedEffortValueAtom } from "@/atoms";
-import { CURRENT, HP } from "@/constants";
+import { CURRENT } from "@/constants";
 import { useErrorToast } from "@/hooks";
-import type { StatusType } from "@/types";
+import type { StatusSpecies, StatusType } from "@/types";
 
 import { Presentation } from "./presentation";
+import { MAX_EFFORT_VALUE, MAX_INDIVIDUAL_VALUE } from "./styleConfig";
 import { useEffortValue } from "../../hooks";
-import { calcHPActualValue } from "../../logic/calcHPActualValue";
-import { calcHPEffortValue } from "../../logic/calcHPEffortValue";
-import { MAX_EFFORT_VALUE, MAX_INDIVIDUAL_VALUE } from "../styleConfig";
+import { calcActualValue } from "../../logic/calcActualValue";
+import { calcEffortValue } from "../../logic/calcEffortValue";
 
-import type { calcActualValue } from "../../logic/calcActualValue";
 import type { MouseEventHandler } from "react";
 
 type Props = {
   level: number;
+  speciesName: StatusSpecies;
   baseStat: number;
   statusType: StatusType;
 };
 
-export function HpStatusTableBody({
+export function StatusTableBody({
+  speciesName,
   baseStat,
   level,
   statusType,
 }: Props) {
   const [individualValue, setIndividualValue] = useState(31);
   const effortValueAtom = statusType === CURRENT ? currentEffortValueAtom : adjustedEffortValueAtom;
+  const [nature, setNature] = useState(1);
 
   const { totalEffortValue, allEffortValue, updateEffortValue } = useEffortValue(effortValueAtom);
 
-  const effortValue = allEffortValue.find((v) => v.name === HP);
+  const effortValue = allEffortValue.find((v) => v.name === speciesName);
 
   const { showErrorToast } = useErrorToast();
   if (!effortValue) {
@@ -40,34 +42,37 @@ export function HpStatusTableBody({
     throw Error();
   }
 
-  const [actualValue, setActualValue] = useState(calcHPActualValue({
+  const [actualValue, setActualValue] = useState(calcActualValue({
     baseStat,
     individual: individualValue,
     effort: effortValue.value,
     level,
+    nature,
   }));
 
-  const minimumActualValue = calcHPActualValue({
+  const minimumActualValue = calcActualValue({
     baseStat,
     individual: individualValue,
     effort: 0,
     level,
+    nature,
   });
 
-  const maximumActualValue = calcHPActualValue({
+  const maximumActualValue = calcActualValue({
     baseStat,
     individual: individualValue,
     effort: 252,
     level,
+    nature,
   });
 
-  const updateActualValue = useCallback((updateValue: Partial<typeof calcActualValue | typeof calcHPActualValue>) => {
-    const newActualValue = calcHPActualValue({
-      baseStat, individual: individualValue, effort: effortValue.value, level, ...updateValue,
+  const updateActualValue = useCallback((updateValue: Partial<typeof calcActualValue>) => {
+    const newActualValue = calcActualValue({
+      baseStat, individual: individualValue, effort: effortValue.value, level, nature, ...updateValue,
     });
 
     setActualValue(newActualValue);
-  }, [baseStat, effortValue.value, individualValue, level]);
+  }, [baseStat, effortValue.value, individualValue, level, nature]);
 
   useEffect(() => {
     updateActualValue(level);
@@ -75,17 +80,17 @@ export function HpStatusTableBody({
 
   const handleChangeActualValue: (_: string, valueAsNumber: number) => void = useCallback((_, value) => {
     setActualValue((value));
-    const newEffortValue = calcHPEffortValue({
-      actual: value, level, baseStat, individual: individualValue,
+    const newEffortValue = calcEffortValue({
+      actual: value, level, baseStat, individual: individualValue, nature,
     });
-    updateEffortValue({ name: HP, value: newEffortValue });
-  }, [baseStat, individualValue, level, updateEffortValue]);
+    updateEffortValue({ name: speciesName, value: newEffortValue });
+  }, [baseStat, individualValue, level, nature, speciesName, updateEffortValue]);
 
   const handleChangeEffortValue: (_: string, valueAsNumber: number) => void = useCallback((_, value) => {
-    updateEffortValue({ name: HP, value });
+    updateEffortValue({ name: speciesName, value });
 
     updateActualValue({ effort: value });
-  }, [updateActualValue, updateEffortValue]);
+  }, [speciesName, updateActualValue, updateEffortValue]);
 
   const maximizeEffortValue: MouseEventHandler<HTMLButtonElement> = useCallback(() => {
     updateEffortValue({
@@ -118,6 +123,11 @@ export function HpStatusTableBody({
     updateActualValue({ individual: 0 });
   }, [updateActualValue]);
 
+  const handleChangeNature: (_: string, valueAsNumber: number) => void = useCallback((_, value) => {
+    setNature(value);
+    updateActualValue({ nature: value });
+  }, [updateActualValue]);
+
   return (
     <Presentation
       actualValue={actualValue}
@@ -125,6 +135,7 @@ export function HpStatusTableBody({
       handleChangeActualValue={handleChangeActualValue}
       handleChangeEffortValue={handleChangeEffortValue}
       handleChangeIndividualValue={handleChangeIndividualValue}
+      handleChangeNature={handleChangeNature}
       individualValue={individualValue}
       maximizeEffortValue={maximizeEffortValue}
       maximizeIndividualValue={maximizeIndividualValue}
@@ -132,7 +143,7 @@ export function HpStatusTableBody({
       minimizeEffortValue={minimizeEffortValue}
       minimizeIndividualValue={minimizeIndividualValue}
       minimumActualValue={minimumActualValue}
-      speciesName={HP}
+      speciesName={speciesName}
       totalEffortValue={totalEffortValue}
     />
   );
