+++ src/features/status/tableRow/components/styleConfig.ts
