+++ src/pages/index.page.tsx
@@ -1,10 +1,11 @@
 import { Box, Center, HStack, Skeleton, Spacer } from "@chakra-ui/react";
 import { Suspense } from "react";
 
+import { CURRENT, ADJUSTED } from "@/constants";
 import { Differences } from "@/features/differences/components";
 import { FormImages } from "@/features/forms/components";
 import { SearchForm } from "@/features/search/components";
-import { StatusTableWrapper } from "@/features/status/components";
+import { Status } from "@/features/status/components";
 
 export function Home() {
   return (
@@ -16,9 +17,20 @@ export function Home() {
         </Suspense>
       </HStack>
       <Center width="100vw">
-        <Suspense fallback={<Skeleton />}>
-          <StatusTableWrapper />
-        </Suspense>
+        <HStack spacing="32px">
+          <Suspense fallback={<Skeleton />}>
+            <Status
+              label="現在のステータス"
+              statusType={CURRENT}
+            />
+          </Suspense>
+          <Suspense fallback={<Skeleton />}>
+            <Status
+              label="調整後のステータス"
+              statusType={ADJUSTED}
+            />
+          </Suspense>
+        </HStack>
       </Center>
       <Box width="100%">
         <Differences />
